@@ -12,52 +12,52 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.2.4",
-    "@angular/common": "^4.2.4",
-    "@angular/compiler": "^4.2.4",
-    "@angular/core": "^4.2.4",
-    "@angular/forms": "^4.2.4",
-    "@angular/http": "^4.2.4",
-    "@angular/platform-browser": "^4.2.4",
-    "@angular/platform-browser-dynamic": "^4.2.4",
-    "@angular/router": "^4.2.4",
-    "@ngrx/effects": "^4.0.2",
-    "@ngrx/store": "^4.0.2",
-    "@ngrx/store-devtools": "^4.0.0",
-    "@swimlane/ngx-datatable": "10.0.5",
-    "@types/lodash-es": "^4.14.6",
-    "@webcomponents/custom-elements": "^1.0.0-alpha.3",
-    "angularfire2": "^4.0.0-rc.1",
-    "clarity-angular": "^0.10.0",
-    "clarity-icons": "^0.10.0",
-    "clarity-ui": "^0.10.0",
-    "core-js": "^2.4.1",
-    "firebase": "^4.2.0",
-    "lodash": "^4.17.4",
-    "lodash-es": "^4.17.4",
-    "rxjs": "^5.4.2",
-    "zone.js": "^0.8.14"
+    "@angular/animations": "6.0.7",
+    "@angular/common": "6.0.7",
+    "@angular/compiler": "6.0.7",
+    "@angular/core": "6.0.7",
+    "@angular/forms": "6.0.7",
+    "@angular/http": "6.0.7",
+    "@angular/platform-browser": "6.0.7",
+    "@angular/platform-browser-dynamic": "6.0.7",
+    "@angular/router": "6.0.7",
+    "@ngrx/effects": "6.0.1",
+    "@ngrx/store": "6.0.1",
+    "@ngrx/store-devtools": "6.0.1",
+    "@swimlane/ngx-datatable": "13.0.1",
+    "@types/lodash-es": "4.17.0",
+    "@webcomponents/custom-elements": "1.1.2",
+    "angularfire2": "5.0.0-rc.11",
+    "@clr/icons": "0.12.3-patch.2",
+    "@clr/ui": "0.12.3-patch.2",
+    "@clr/angular": "0.12.3-patch.2",
+    "core-js": "2.5.7",
+    "firebase": "5.2.0",
+    "lodash-es": "4.17.10",
+    "rxjs": "6.2.1",
+    "zone.js": "0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.3.0",
-    "@angular/compiler-cli": "^4.2.4",
-    "@angular/language-service": "^4.2.4",
-    "@types/jasmine": "~2.5.53",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/lodash": "^4.14.73",
-    "@types/node": "~6.0.60",
-    "codelyzer": "~3.1.1",
-    "jasmine-core": "~2.6.2",
-    "jasmine-spec-reporter": "~4.1.0",
-    "karma": "~1.7.0",
-    "karma-chrome-launcher": "~2.1.1",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
-    "ts-node": "~3.2.0",
-    "tslint": "~5.3.2",
-    "typescript": "~2.3.3"
+    "@angular-devkit/build-angular": "~0.6.8",
+    "@angular/cli": "6.0.8",
+    "@angular/compiler-cli": "6.0.7",
+    "@angular/language-service": "6.0.7",
+    "@types/jasmine": "2.8.8",
+    "@types/jasminewd2": "2.0.3",
+    "@types/lodash": "4.14.110",
+    "@types/node": "10.5.2",
+    "codelyzer": "4.4.2",
+    "jasmine-core": "3.1.0",
+    "jasmine-spec-reporter": "4.2.1",
+    "karma": "2.0.4",
+    "karma-chrome-launcher": "2.2.0",
+    "karma-cli": "1.0.1",
+    "karma-coverage-istanbul-reporter": "2.0.1",
+    "karma-jasmine": "1.1.2",
+    "karma-jasmine-html-reporter": "1.2.0",
+    "protractor": "5.3.2",
+    "ts-node": "7.0.0",
+    "tslint": "5.10.0",
+    "typescript": "^2.7.2"
   }
 }
