@@ -1,11 +1,12 @@
 import { Injectable } from '@angular/core';
 import { AngularFireDatabase } from 'angularfire2/database';
-import { Observable } from 'rxjs/Observable';
+import { Observable, concat, from } from 'rxjs';
 
 import { IPick } from './IPick';
 import { IPlayer } from '../players/IPlayer';
 import { OwnersDAO } from '../owners/ownersDAO';
 import { PlayersDAO } from '../players/playersDAO';
+import { reduce } from "rxjs/operators";
 
 @Injectable()
 export class PicksDAO {
@@ -20,19 +21,19 @@ export class PicksDAO {
   ) {}
 
   getPicks(): Observable<IPick[]> {
-    return this.firebase.list(this.picksUrl);
+    return this.firebase.list<IPick>(this.picksUrl).valueChanges();
   }
 
   getCurrentPick(): Observable<IPick> {
-    return this.firebase.object(this.currentPickUrl);
+    return this.firebase.object<IPick>(this.currentPickUrl).valueChanges();
   }
 
   getPreviousPick(): Observable<IPick> {
-    return this.firebase.object(this.previousPickUrl);
+    return this.firebase.object<IPick>(this.previousPickUrl).valueChanges();
   }
 
   selectPlayer(pick: IPick, player: IPlayer, nextPick: IPick): Observable<any> {
-    return Observable.concat(
+    return concat(
       // set picks
       this.firebase.object(`${this.picksUrl}/${pick.overallSelection}/player`).set(player),
       // update player is selected flag
@@ -43,16 +44,17 @@ export class PicksDAO {
       this.firebase.object(this.currentPickUrl).set(nextPick),
       // set the previous pick to the current pick
       this.firebase.object(this.previousPickUrl).set(pick)
-    )
-      .reduce(() => {});
+    ).pipe(
+      reduce(() => {})
+    );
   }
 
   changeCurrentPick(newPick: IPick): Observable<any> {
-    return Observable.fromPromise(this.firebase.object(this.currentPickUrl).set(newPick));
+    return from(this.firebase.object<IPick>(this.currentPickUrl).set(newPick));
   }
 
   unselectPlayer(pick: IPick, player: IPlayer): Observable<any> {
-    return Observable.concat(
+    return concat(
       // remove from /pick/:id
       this.firebase.object(`${this.picksUrl}/${pick.overallSelection}/player`).remove(),
       // update player is taken flag
@@ -62,7 +64,8 @@ export class PicksDAO {
       // set current pick to the removal pick,
       this.firebase.object(this.currentPickUrl).set(_.omit(pick, ['player']))
       // don't think i need to update previous pick
-    )
-      .reduce(() => {});
+    ).pipe(
+      reduce(() => {})
+    );
   }
 }
