@@ -1,14 +1,15 @@
-import { Component, OnInit } from '@angular/core';
-
-import { Store } from '@ngrx/store';
+import { Component, OnDestroy, OnInit } from '@angular/core';
 import { ActivatedRoute, Params } from '@angular/router';
+
+import { select, Store } from '@ngrx/store';
+import { Observable, Subject } from 'rxjs';
+import { tap, publishReplay, refCount, takeUntil } from "rxjs/operators";
 
 import { IAppState } from './state/appState';
 import {
   GetCurrentPickAction, GetOwnersAction, GetPicksAction, GetPlayersAction,
   MakeAdminAction
 } from './state/appActions';
-import { Observable } from 'rxjs/Observable';
 
 @Component({
   moduleId: module.id,
@@ -16,31 +17,39 @@ import { Observable } from 'rxjs/Observable';
   template: `<app-root [state]="state$ | async"></app-root>`
 })
 
-export class AppContainerComponent implements OnInit {
+export class AppContainerComponent implements OnInit, OnDestroy {
   state$: Observable<IAppState>;
+  ngDestroy$ = new Subject<any>();
 
   constructor(
     private store: Store<{ app: IAppState }>,
     private route: ActivatedRoute
   ) {
-    this.state$ = this.store
-      .select(thing => thing.app)
-      .publishReplay(1)
-      .refCount();
+    this.state$ = this.store.pipe(
+      select(thing => thing.app),
+      publishReplay(1),
+      refCount()
+    );
   }
 
   ngOnInit(): void {
-    this.route.queryParams
-      .do((params: Params) => {
+    this.route.queryParams.pipe(
+      tap((params: Params) => {
         if (params.admin === 'true') {
           this.store.dispatch(<MakeAdminAction> { type: 'MAKE_USER_ADMIN' });
         }
-      })
-      .subscribe();
+      }),
+      takeUntil(this.ngDestroy$)
+    ).subscribe();
 
     this.store.dispatch(<GetOwnersAction> { type: 'GET_OWNERS' });
     this.store.dispatch(<GetPicksAction> { type: 'GET_PICKS' });
     this.store.dispatch(<GetPlayersAction> { type: 'GET_PLAYERS' });
     this.store.dispatch(<GetCurrentPickAction> { type: 'GET_CURRENT_PICK' });
+  }
+
+  ngOnDestroy(): void {
+    this.ngDestroy$.next();
+    this.ngDestroy$.complete();
   }
 }
